// ui/src/shared/types/permission.types.ts
export enum Permission {
  VIEW_USERS = 'view:users',
  MANAGE_USERS = 'manage:users',
  VIEW_ROLES = 'view:roles',
  MANAGE_ROLES = 'manage:roles',
  VIEW_PROFIT = 'view:profit',
  VIEW_BLACKLISTS = 'view:blacklists',
  MANAGE_BLACKLISTS = 'manage:blacklists',
  EXEC_SCHEDULE_MARKET_SIGNAL = 'exec:schedule:market_signal',
  EXEC_SCHEDULE_ALLOCATION_RECOMMENDATION_NEW = 'exec:schedule:allocation_recommendation_new',
  EXEC_SCHEDULE_ALLOCATION_RECOMMENDATION_EXISTING = 'exec:schedule:allocation_recommendation_existing',
  EXEC_SCHEDULE_ALLOCATION_AUDIT = 'exec:schedule:allocation_audit',
  VIEW_INFERENCE_NASDAQ = 'view:inference:nasdaq',
  VIEW_INFERENCE_COIN_MAJOR = 'view:inference:coin:major',
  VIEW_INFERENCE_COIN_MINOR = 'view:inference:coin:minor',
  TRADE_NASDAQ = 'trade:nasdaq',
  TRADE_COIN_MAJOR = 'trade:coin:major',
  TRADE_COIN_MINOR = 'trade:coin:minor',
}